refact burger ingredients
