+++ src/components/burger-ingredients/burger-items/burger-items.module.css
@@ -1,0 +1,32 @@
+.productList {
+  overflow-y: scroll;
+  height: 500px;
+  scrollbar-color: #696969 #8585ad;
+  scrollbar-width: thin;
+}
+
+.productList::-webkit-scrollbar {
+  width: 12px;
+}
+.productList::-webkit-scrollbar-track {
+  background: #2f2f37;
+  margin-top: 40px;
+}
+.productList::-webkit-scrollbar-thumb {
+  background: #8585ad;
+  border: 3px solid #2f2f37;
+  border-radius: 6px;
+}
+
+.ingredients {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-around;
+}
+
+@media screen and (min-height: 768px) {
+  .productList {
+    height: 50vh;
+  }
+}
